feat: implement core bot logic and event handling in bot module
